requirements.txt
python-telegram-bot>=20.0
icmplib>=3.0
schedule>=1.2.0
python-dotenv>=1.0.0
httpx>=0.24.0